requirements.txt
fastapi
uvicorn
openai
pytesseract
Pillow
faiss-cpu
sentence-transformers
beautifulsoup4
requests
